Allow updates when release infor change for (#25)
https://manpages.debian.org/unstable/apt/apt-get.8.en.html
> --allow-releaseinfo-change

Allow the update command to continue downloading data from a repository which changed its information of the release contained in the repository indicating e.g a new major release. APT will fail at the update command for such repositories until the change is confirmed to ensure the user is prepared for the change. See also apt-secure(8) for details on the concept and configuration.
Specialist options (--allow-releaseinfo-change-field) exist to allow changes only for certain fields like origin, label, codename, suite, version and defaultpin.

Co-authored-by: Harsh Parekh <harsh.parekh@verbasoftware.com>