Corrected installation location of chromedriver binary (#81)
In previous chromedriver releases, the ZIP archives would contain the
'chromedriver' binary in the root directory of the archive. This file
was then moved to the target directory, typically

  /usr/local/bin/chromedriver

With chromedriver 15 however, the ZIP archives contain a new
platform-specific subdirectory. Thus, the actual binary is at e.g.
'chromedriver-linux64/chromedriver'. As it is, the
'chromedriver-linux64' directory is renamed to 'chromedriver' and *that*
is then moved to the target directory. This results in the final
executable to end up at

  /usr/local/bin/chromedriver/chromedriver

This is not backwards compatible with usages of the orb expecting the
previous location (e.g. because they expect that 'install-chromedriver'
puts the binary into a directory which is part of the PATH).